Updated to two versions of the tasks
