+++ src/swagger-diff/swagger-diff-v1/icon.png
