+++ src/swagger-diff/swagger-diff-v1/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "vsts-swagger-diff",
+  "version": "0.0.1",
+  "description": "A build/release task for comparing Swagger Contracts.",
+  "homepage": "",
+  "author": {
+    "name": "knom"
+  },
+  "files": [],
+  "main": "",
+  "keywords": [],
+  "dependencies": {
+    "azure-pipelines-task-lib": "^3.1.0",
+    "growl": ">=1.10.5",
+    "lodash": ">=4.17.21",
+    "node-fetch": "^2.6.1",
+    "swagger-diff": "^0.6.0"
+  }
+}
