+++ src/swagger-diff/swagger-diff-v1/task.json
@@ -1,0 +1,134 @@
+{
+  "id": "d0b32dd5-f8bd-4118-8e66-699ad33189a2",
+  "name": "swagger-diff",
+  "friendlyName": "Swagger Contract Diff",
+  "description": "A build/release task for comparing two Swagger Contracts.",
+  "helpMarkdown": "### Compare two Swagger Contracts - left-side and right-side.",
+  "category": "Test",
+  "visibility": [
+    "Build",
+    "Release"
+  ],
+  "author": "knom",
+  "version": {
+    "Major": 1,
+    "Minor": 0,
+    "Patch": 2
+  },
+  "demands": [
+    "npm"
+  ],
+  "groups": [
+    {
+      "name": "left",
+      "displayName": "Left-Side Document",
+      "isExpanded": true
+    },
+    {
+      "name": "right",
+      "displayName": "Right-Side Document",
+      "isExpanded": true
+    },
+    {
+      "name": "advanced",
+      "displayName": "Advanced",
+      "isExpanded": false
+    }
+  ],
+  "inputs": [
+    {
+      "name": "swaggerVersion",
+      "type": "radio",
+      "label": "Swagger Version:",
+      "defaultValue": "v2",
+      "required": "true",
+      "helpMarkDown": "Which version of Swagger is being used",
+      "options": {
+        "v2": "Swagger v2",
+        "v3": "OpenAPI v3"
+      }
+    },
+    {
+      "name": "leftType",
+      "type": "radio",
+      "label": "Left Document Type:",
+      "defaultValue": "file",
+      "required": "true",
+      "helpMarkDown": "Whether the left-side Swagger document is a file path or Url.",
+      "options": {
+        "file": "File",
+        "url": "Url"
+      },
+      "groupName": "left"
+    },
+    {
+      "name": "leftFilePath",
+      "type": "filePath",
+      "label": "Left Document File Path:",
+      "defaultValue": "",
+      "required": "true",
+      "helpMarkDown": "The file path of the left-side Swagger document.",
+      "visibleRule": "leftType = file",
+      "groupName": "left"
+    },
+    {
+      "name": "leftUrl",
+      "type": "string",
+      "label": "Left Document Url:",
+      "defaultValue": "",
+      "required": "true",
+      "helpMarkDown": "The URL of the left-side Swagger document.",
+      "visibleRule": "leftType = url",
+      "groupName": "left"
+    },
+    {
+      "name": "rightType",
+      "type": "radio",
+      "label": "Right Document Type:",
+      "defaultValue": "file",
+      "required": "true",
+      "helpMarkDown": "Whether the right-side Swagger document is a file path or Url.",
+      "options": {
+        "file": "File",
+        "url": "Url"
+      },
+      "groupName": "right"
+    },
+    {
+      "name": "rightFilePath",
+      "type": "filePath",
+      "label": "Right Document File Path:",
+      "defaultValue": "",
+      "required": "true",
+      "helpMarkDown": "The file path of the right-side Swagger document.",
+      "visibleRule": "rightType = file",
+      "groupName": "right"
+    },
+    {
+      "name": "rightUrl",
+      "type": "string",
+      "label": "Right Document Url:",
+      "defaultValue": "",
+      "required": "true",
+      "helpMarkDown": "The URL of the right-side Swagger document.",
+      "visibleRule": "rightType = url",
+      "groupName": "right"
+    },
+    {
+      "name": "config",
+      "type": "string",
+      "label": "Swagger-Diff Configuration Override:",
+      "defaultValue": "",
+      "required": "false",
+      "helpMarkDown": "The optional JSON Configuration override for swagger-diff.",
+      "groupName": "advanced"
+    }
+  ],
+  "instanceNameFormat": "Swagger Contract Diff $(name)",
+  "execution": {
+    "Node10": {
+      "target": "task.js",
+      "argumentFormat": ""
+    }
+  }
+}
