+++ src/swagger-diff/swagger-diff-v2/.eslintrc.json
@@ -1,0 +1,8 @@
+{
+    "parserOptions": {
+        "ecmaVersion": 2020
+    },
+    "rules": {
+        "no-unused-vars": "error"
+    }
+}
